Add github.com/gibizer/operator-lint
To be able to add that in the same way as in other service-operators the
existing operator-lint target is renamed to kube-linter and
operator-lint target now run github.com/gibizer/operator-lint.

The kube-linter target extended to ensure that the kube-linter binary is
installed locally.

The operator-lint need go.work setup to be able to check the ./api module
properly so a gowork make target is added and go.work is ignored in
.gitignore.

The go work sync found a conflict in dependencies between the
main and the ./api module so github.com/kr/pretty is bumped to v0.2.1 to
resolve that.

Then the operator-lint findings are fixed.