Rewrite telemetry resource
Rewrite the telemetry resource to use a similar pattern to the
openstack control plane resource. The telemetry resource is an
umbrela resource for other telemetry related resources, which
for right now are "ceilometer" and "autoscaling". These resources
can be enabled and customized by using the "enabled" and
"template" fields similarly to how openstack services are
enabled and customized in the openstack control plane CR.

The "secret" field in the telemetry resource can be used
to set the "secret" fields for both ceilometer and autoscaling.

The autoscaling is disabled by default and ceilometer
is enabled by default.

Look at the config/samples/telemetry_v1beta1_telemetry.yaml for
an example CR.