Use PodManagementPolicy: appsv1.ParallelPodManagement for the aodh/ceilometer statefulset
With the default `PodManagementPolicy: OrderedReadyPodManagement` the
statefulset controller will only progress pods when the
previous/current pod is ready or terminated.

When service configuration changes while the pod is starting and the
new configuration requires e.g. additional volume mounts the initial
pod will never reach ready and therefore an update won't happen.

With ParallelPodManagement the statefulset controller will not wait
for pods to be ready or complete termination.