[OSPRH-8065] Rework custom prometheus CA
Rework how the user specifies CA certificates for
user deployed / external prometheus.

Before the PR the user deployed
prometheus was almost unusable. The TLS was always used
no matter the settings (which was a bug). The operator also assumed
that the CA cert is a part of the openstack ca bundle.
This assumption doesn't make sense for external prometheus.

With this PR the user has an option to specify a secret name
and key within that secret which holds the CA certificate.

The functionality around using the metric-storage prometheus
by default is still unchanged and works with and without
controlplane TLS as before.