[OSPRH-8065] Change autoscaling TLS condition
Change the condition for which Prometheus to use for autoscaling.

I noticed the status of autoscaling wasn't being set to Ready
when trying to use user-deployed prometheus while having the
MetricStorage disabled. It'd wrongly try to get a MetricStorage, which
wasn't there and set the error into the status.

Now we're using a better condition. We're looking for MetricStorages
only when we're using a MetricStorage Prometheus. When using
a user-deployed Prometheus, we aren't looking at MetricStorage.