[OSPRH-8406] Switch to using ScrapeConfigs
We recently discovered issues with authentication, IPv6
and ServiceMonitors in STF. This PR is proactively switching
to use ScrapeConfigs instead of ServiceMonitors. The
functinality should be equivalent to before. Old ServiceMonitors
owned by the MetricStorage controller are deleted.

There is a slight difference in the labels associated with
the collected metrics.
 - The Node Exporter metrics are now missing the "job" label,
   which didn't seem useful and it follows how ceilometer
   and rabbit metrics are collected.
 - Ceilometer and RabbitMQ metrics don't have the "service"
   label anymore, because ScrapeConfigs don't have the information
   to create that label. Instead they now have the "instance"
   label.

The "instance" label is now used to differentiate between different
Rabbit clusters in dashboards instead of the "service" label.

I used this opportunity to move the ScrapeConfig creation code
into its own function, following the example of dashboard code.