Lower autoscaling alarm threshold to 50%
We see in the CI, that sometimes the CPU metric doesn't reach
above 60%, so the stack never scales up and the CI fails.
I've never seen it stop below 50%, so this threshold would always
be reached in the past.

The core issue of such a low values are possibly caused
by the length of the load we create in the autoscaled instances.
The load is created only for 2 minutes, which is too low
when ceilometer polls it only every 2 minutes. With bad
luck in timing we will get these low metrics. I'll also increase
the load length in tempest.