Do not ignore err on GetSecret()
In CI jobs it was seen that the controller menager restarted due
to a nil ptr when set the transport url in the config. The issue
iss that when the TransportURL secret gets fetched, no check is
perforned if an error happened.

This adds the err checks on the TransportURL and Password secret.

Signed-off-by: Martin Schuppert <mschuppert@redhat.com>