Add ksmCaBundleSecretNameField and ksmTLSField to the index
64fcf619454865fba54ee5c7974b9142d1b9f017 introduced TLS to KSM
but missed to add the TLS fields to the index. As a result when
the ceilometer cert secret changes, the Watches will get that
the resource changed:

2024-10-10T11:09:17Z    INFO    Controllers.Ceilometer  input source cert-ceilometer-internal-svc changed, reconcile: ceilometer - openstack
2024-10-10T11:09:18Z    INFO    Controllers.Ceilometer  input source cert-ceilometer-internal-svc changed, reconcile: ceilometer - openstack

With this the findObjectsForSrc() exited with an hidden error:
~~~
"error": "Index with name field:.spec.ksmTls.caBundleSecretName does not exist"
~~~

This adds the fields to the index and also updates to log the error.

Signed-off-by: Martin Schuppert <mschuppert@redhat.com>