Fix pagination links returned in non-TLS deployments
Server returns links to the "next" and "previous" page of
results when pagination is requested by client.
In the httpd container configuration there was an issue that
"X-Forwarded-Proto" header was setting "https" protocol always, no
matter if TLS was enabled in the deployment or not. It caused issues
when TLS was disabled as client was doing request to the "http://"
endoint and got in response "https://" links to "next" and "prev" pages.

This patch fixes this issue by setting correct protocol for both cases:
TLS and non-TLS.

Related: OSPRH-10632

Signed-off-by: Martin Schuppert <mschuppert@redhat.com>