[OSPRH-10800] Watch MetricStorage from Autoscaling
MetricStorage is our main source for autoscaling metrics. We
default to MetricStorage if other Prometheus instance isn't
explicitly configured, so we need to ensure that Autoscaling is configured
correctly to access this MetricStorage.

We previously decided to not enforce an order of deployment
of telemetry resources, so it's pretty valid to first deploy
Autoscaling and only after that deploy a MetricStorage. But
if MetricStorage isn't deployed, Autoscaling doesn't know if
it should configure Aodh to use TLS for Prometheus access or not.
Without watching the MetricStorage resources, Autoscaling doesn't
know if a MetricStorage CR is created or changed. So it's
possible to do the following:
 - Enable Autoscaling (it gets automatically configured to not use TLS
   for Prometheus)
 - Enable MetricStorage with TLS enabled
 - Aodh can't access Prometheus because of TLS issues.

With this PR Autoscaling reacts to MetricStorage changes and so
Aodh gets always correctly configured.