enhance NodeExporter metrics
This adds a "hostname" label to NodeExporter metrics
to make the identification of the node of origin a little easier
than just having the IP address.

As a side effect of needing a little different parsing for
NodeExporter and Kepler ScrapeConfigs, the code to get compute
node IPs from ansible facts and from IPsets is now executed
only once (in comparison to executing it twice), which greatly
reduces k8s api calls.