[OSPRH-12359] Fix getting IPSets
Turns out IPSets use the node's hostname as defined in the
nodeset CR, not the node's name. These 2 might be the same, but
they might differ and in case they differ, the controller
doesn't find them and it uses the ansible inventory secret to get
the IPs instead.

Unfortunatelly the hostname from the nodeset isn't always
copied into the ansible inventory secret. It either gets
copied into the canonical_hostname variable as is or
a DNS domain is appended to it, so I introduced a bit
of code to pick the correct parts of the canonical_hostname.

This worked for both cases described above when I tested it,
there is also still the fallback to using the ansibleHost
from the inventory secret.