Treat empty *Image as the feature being disabled
This adds a behavior already present for ksm to mysqldExporter.
After conversation with Martin we decided to modify ksm as well
and output a log if a feature is enabled, but the required
image isn't set. We can't set a status condition (at least not for now)
as defaulting the image isn't reliable and then it blocks the
whole controlplane.