Add openstack-approvers to approvers/reviewers
The source for ci-operator/config/openstack-k8s-operators/telemetry-operator/OWNERS
in the the CI configuration at github.com/openshift/release is the
OWNERS of this repo.
To update the CI configuration for e.g. feature branching, the
openstack-approvers should be in the OWNERS_ALIASES/OWNERS
like for in the other repositories.

Signed-off-by: Martin Schuppert <mschuppert@redhat.com>