[ci] Increase timeout in COO install
Sometimes, COO fails to become available in the alotted time.
This can be caused by resource contention or slow networks.
This commit increases the number of retries to give the resources
up to 10 minutes to appear.

This will not effect the runtimes of otherwise healthy builds,
however, it should reduce the amount of failures associated with
failure of this task.
Furthermore, it will allow us to check the number of times the job
would have failed at this tassk, by checking the number of retries
actually used.

JIRA: OSPRH-16776