[DNM][ci] add temp config
This is to workaround a previous workaround (revert is in progress)
This change should not be merged, but is intended to unblock CI while we wait
for the fixes to be merged into the operators.

Since https://review.opendev.org/c/openstack/oslo.cache/+/950561 was
merged, the workaournd is not needed, but the reverts have not all
merged yet.