[CloudKitty] Add osp-secret as the default for secret
When the secret is unset, cloudkitty will not start, since there was no
default set.

The osp-secret is used as the default secret for other services.

Thes default value for the cloudkitty secret is updated to match the
other service defaults, since the required information (CK password)
is expected to be in the same secret as the other services.