[cloudkitty] Update the cloudkitty.conf to reduce the number of workers
The numbers are now set really low, but when they're unset, they are
set to maximum.

The default values for both the rating process and the reprocessing
is the number of CPUs.
The default values for the threads per worker is 5x the total number
of CPUs.

This means that cloudkitty can potentially tie up all the CPUs on the
hosts, since the pods don't necessarily have limits on the resources
they have access to.

The updated values are up to 2 workers each for rating and reprocessing,
and up to 2 threads per worker.