[cloudkitty] Update probes and mounts for CloudKitty
The healthcheck for cloudkitty-proc/probe was failing because the CA
was not trusted. The TLS-related file were not available in the expected
location in the container.

The kolla_set_config command is used so that the mounted config files
get copied into the expected location for the healthcheck.

The command was updated to use 'bash -c' so that both the kolla_set_config
and the healthcheck command would be run on startup.

[API] Update script used in uwsgi template

The script location changed between antelope and master containers.
In both cases, it exists in the same alternative location

Update initial delay on liveness and readiness probes
This give a chance for the api to start up before the first healthcheck

[API] Add metrics.yaml into the api pod