[cloudkitty] Add a default value for cloudkitty.s3StorageConfig
The s3StorageConfig.secret is passed to Lokistack, which requires
some value be set.
The default is needed, or else there will be an error when it is not
configured.

The default is set to a reasonable value for the secret name
(cloudkitty-loki-s3) and the type (s3), which match the values we're
planning on using for the CI environment.

Lines added: 1
Lines generated: 16