[ci] Force metacontent provider to always rebuild the container images
The current-tested tag expired from the openstack-ceilometer-compute
image on quay.rdoproject.org/podified-master-centos10, which causes the
edpm deployment to fail, since the image is not available.

This is being temperorily worked around by forcing the meta-content
provider to always build and use all the telemetry images.

This is not an ideal solution, since the image builds take time, and
this increases the overall buildset duration.