Create a NAD in the Kuttl namespace
The ctlplane NAD exists in the openstack namespace
https://github.com/openstack-k8s-operators/install_yamls/blob/ba562f76c944c4bc92f7a4efe092e7ec30aebce4/scripts/gen-netatt.sh#L71
and this might be the reason the required NAD is not
being picked up during the metricstorage creation

Try precreating the NAD. Maybe some other resources are needed,
will follow up if that is the case.

Also, get back to updating the existing metricstorage so we
can test that updating the metricstorage when TLS is enabled
doesn't fall in a reconciliation loop.