[OSPRH-21020] Scrape all rabbitmq pods
RabbitMQ can be deployed with multiple replicas, while each replica
will export different metrics. Having a ScrapeConfig with a service
as a target means we're not scraping all metrics. This PR
uses the Endpoints object to retrieve IP addresses of all rabbitmq
related pods and creates a ScrapeConfig to scrape each pod through
its IP address. This keeps the pattern of using ScrapeConfigs for
configuring scraping.