Workaround for kuttl tests for CI, needs follow up
telemetry-operator is special in the kuttl config compared to others.
it defined test dirs in kuttl-test.yaml and uses an sed in install_yamls
to change it if needed.
Other operators pass an updated test path to the kubectl-kutt command in
install_yamls.

Because of that we
* keep the old tests path in kuttl-test.yaml.
* use a link tests -> test dir

When a fix landed in install_yamls to be the same as other operators,
we remove the link and update the kuttl-test.yaml.

Signed-off-by: Martin Schuppert <mschuppert@redhat.com>