Fix tls kuttl-test suite
COO 1.3.1 released recently, which increased how many rulefiles
volumeMounts the prometheus pod has. We're checking that the tls
certificates are being mounted into the prometheus pod and so
this change in count of volumeMounts resulted in the kuttl-tests
failing with:
.spec.containers.volumeMounts: slice length mismatch: 9 != 11

This PR adds the 2 additional volume mounts in the kuttl-test
assert.